pyproject.toml
[project]
name = "docpond-core"
version = "0.1.0"
description = "Document management backend for scanned books with page-level search"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core Django dependencies for both containers
    "django>=6.0.1",
    "django-sqids>=0.2.2",
    "pydantic>=2.0.0",
    "redis>=5.0.0",
    "django-ninja>=1.3.0",
    "psycopg2-binary>=2.9.0",
    "pillow>=10.0.0",
    "celery>=5.3.0",
    "pymupdf>=1.24.0",
    "django-cors-headers>=4.9.0",
    "django-storages>=1.14.0",
    "boto3>=1.26.0",
    "django-environ>=0.12.0",
    "dj-database-url>=3.1.0",
    "docling>=2.69.1",
    "ocrmypdf>=16.0.0",
]
